improvements after review
