+++ src/redisearch_rs/trie_bencher/src/main.rs
@@ -2,7 +2,8 @@ use std::{collections::BTreeSet, ptr::NonNull};
 use trie_bencher::corpus::CorpusType;
 use trie_bencher::{
     CTrieMap, RustTrieMap,
-    ffi::{AsTrieTermView as _, ToCstr as _, str2boxed_c_char},
+    cmap::{AsTrieTermView as _, IntoCstring as _},
+    str2boxed_c_char,
 };
 
 fn main() {
@@ -35,7 +36,7 @@ fn compute_and_report_memory_usage() {
             map.insert(&converted, value);
 
             // C insertion
-            cmap.insert(&word.to_cstr().as_view());
+            cmap.insert(word.into_cstring().as_view());
         }
     }
 
@@ -48,7 +49,7 @@ fn compute_and_report_memory_usage() {
         );
         assert!(
             // Safety: TRIEMAP_NOTFOUND is a constant defined in the C code
-            cmap.find(&unique_word.to_cstr().as_view())
+            cmap.find(unique_word.into_cstring().as_view())
                 != unsafe { trie_bencher::ffi::TRIEMAP_NOTFOUND },
             "{unique_word} not found in C map"
         )
